getting_data/pubmed_scraping/requirements.txt
biopython
html2text